zad11

+++ JavaBookNakov/src/chapter06/Zad11.java
@@ -6,11 +6,11 @@ public class Zad11 {
 	public static void main(String[] args) {
 		Scanner scan = new Scanner(System.in);
 		System.out.print("Enter the K!");
-		int K = scan.nextInt();
+		long K = scan.nextLong();
 
 		//Faktoriel
-		int sum = K;
-		for (int i = K; i > 1; i--) {
+		long sum = K;
+		for (long i = K; i > 1; i--) {
 			sum *= ((i - 1));
 		}
 		K = sum;
@@ -18,7 +18,7 @@ public class Zad11 {
 
 	
 		int count = 0;
-		for (int tmp = K; tmp % 5 == 0; tmp /= 5) {
+		for (long tmp = K; tmp % 5 == 0; tmp /= 5) {
 
 			count++;
